Update requirements algorithm in menu
