+++ src/components/Requirements.vue
@@ -18,14 +18,14 @@
       <div class="progress">
         <div
           class="progress-bar"
-          v-bind:style="{ 'background-color': `#${req.color}`, width: `${(req.progress/req.total)*100}%`}"
+          v-bind:style="{ 'background-color': `#${req.color}`, width: `${(req.fulfilled/req.required)*100}%`}"
           role="progressbar"
         ></div>
       </div>
 
       <p class="progress-text">
-        <strong>{{ req.progress }}/{{ req.total }}</strong>
-        Total {{ req.count }} Inputted on Schedule
+        <strong>{{ req.fulfilled }}/{{ req.required }}</strong>
+        Total {{ req.type }} Inputted on Schedule
       </p>
 
       <!--View more college requirements -->
@@ -42,7 +42,7 @@
           </button>
         </div>
          <div class="col p-0 ">
-          <p class="req-name" v-bind:style="{ 'color': `#${req.color}` }">{{ (req.displayDetails) ? "HIDE" : "VIEW" }} ALL {{ req.type }} REQUIREMENTS</p>
+          <p class="req-name" v-bind:style="{ 'color': `#${req.color}` }">{{ (req.displayDetails) ? "HIDE" : "VIEW" }} ALL {{ req.group }} REQUIREMENTS</p>
         </div>
       </div>
 
@@ -69,7 +69,7 @@
               <p class="sup-req">{{subReq.name}}</p>
             </div>
             <div class="col">
-              <p class="sup-req-progress text-right">( {{subReq.progress}}/{{subReq.total}} Credits)</p>
+              <p class="sup-req-progress text-right">( {{subReq.fulfilled}}/{{subReq.required}} Credits )</p>
             </div>
           </div>
 
@@ -140,7 +140,7 @@
                 <p class="sup-req">{{subReq.name}}</p>
               </div>
               <div class="col">
-                <p class="sup-req-progress text-right">( {{subReq.progress}}/{{subReq.total}} Credits)</p>
+                <p class="sup-req-progress text-right">( {{subReq.fulfilled}}/{{subReq.required}} Credits )</p>
               </div>
             </div>
           </div>
@@ -159,7 +159,7 @@ import VueCollapse from 'vue2-collapse';
 import Course from '@/components/Course';
 import Modal from '@/components/Modals/Modal';
 import Semester from '@/components/Semester';
-// const fs = require('fs');
+
 import reqsData from '../requirements/reqs.json';
 
 const request = require('request');
@@ -180,180 +180,460 @@ export default {
     compact: Boolean
   },
   mounted() {
-    this.getReqs();
+    this.getReqs(["CS 2110", "CS 1110", "PE 1110", "PE 1300"], "AS", "CS").then(groups => {
+      // Turn result into data readable by requirements menu
+      console.log(groups);
+
+      groups.forEach(group => {
+        let singleMenuRequirement = {ongoing: [], completed: []};
+        singleMenuRequirement.name = `${group.groupName.toUpperCase()} REQUIREMENT`;
+        singleMenuRequirement.group = group.groupName.toUpperCase();
+        singleMenuRequirement.color = '105351';
+        singleMenuRequirement.displayDetails = false;
+        singleMenuRequirement.displayCompleted = false;
+        
+        group.reqs.forEach(req => {
+          if (req.progressBar) {
+            singleMenuRequirement.type = req.type.charAt(0).toUpperCase() + req.type.substring(1);
+            singleMenuRequirement.fulfilled = req.fulfilled;
+            singleMenuRequirement.required = req.required;
+          }
+
+          // Default display value of false for all requirement lists
+          req.display = false;
+
+          if (!req.fulfilled || req.fulfilled < req.required) {
+            singleMenuRequirement.ongoing.push(req);
+          } else {
+            singleMenuRequirement.completed.push(req);
+          }
+        })
+
+        if (!singleMenuRequirement.type) {
+          singleMenuRequirement.type = "Requirements"
+          singleMenuRequirement.fulfilled = singleMenuRequirement.completed.length;
+          singleMenuRequirement.required = singleMenuRequirement.ongoing.length + singleMenuRequirement.completed.length;
+        }
+
+        this.reqs.push(singleMenuRequirement);
+      })
+    });
   },
 
   data() {
-    const randomId = Math.floor(Math.random() * Math.floor(100));
     return {
       reqsData,
       actives: [false],
       modalShow: false,
-      reqs1: [],
       reqs: [
-        {
-          name: 'COLLEGE REQUIREMENTS',
-          type: 'COLLEGE',
-          count: 'Credits',
-          progress: 46,
-          total: 120,
-          color: '2BBCC6',
-          displayDetails: false,
-          ongoing: [
-            {
-              name: 'CALS Credits',
-              count: 'Credits',
-              progress: 12,
-              total: 55,
-              additonalCourses: ['PE Courses', 'Swimming Requirement'],
-              satisfiableCourses: ['ice skating', 'bowling'],
-              display: false
-            },
-            {
-              name: 'PE Credits',
-              count: 'Credits',
-              progress: 1,
-              total: 2,
-              additonalCourses: ['PE Courses', 'Swimming Requirement'],
-              satisfiableCourses: ['ice skating', 'bowling'],
-              display: false
-            }
-          ],
-          displayCompleted: true,
-          completed: [
-            {
-              name: 'Quantitative Literacy',
-              count: 'Credits',
-              progress: 2,
-              total: 2,
-              display: false
-            },
-            {
-              name: 'Chemistry/Physics',
-              count: 'Credits',
-              progress: 1,
-              total: 2,
-              display: false
-            }
-          ]
-        },
-        {
-          name: 'MAJOR REQUIREMENTS',
-          type: 'MAJOR',
-          count: 'Courses',
-          progress: 4,
-          total: 8,
-          color: '508197',
-          displayDetails: false,
-          ongoing: [
-            {
-              name: 'CALS Credits',
-              count: 'Credits',
-              progress: 12,
-              total: 55,
-              display: false
-            },
-            {
-              name: 'PE ',
-              count: 'Credits',
-              progress: 1,
-              total: 2,
-              display: false
-            }
-          ],
-          displayCompleted: true,
-          completed: [
-            {
-              name: 'Quantitative Literacy',
-              count: 'Credits',
-              progress: 2,
-              total: 2,
-              display: false
-            },
-            {
-              name: 'Chemistry/Physics',
-              count: 'Credits',
-              progress: 2,
-              total: 2,
-              display: false
-            }
-          ]
-        },
+        // {
+        //   name: 'UNIVERSITY REQUIREMENT',
+        //   group: 'UNIVERSITY',
+        //   type: 'Credits',
+        //   fulfilled: 46,
+        //   required: 120,
+        //   color: '105351',
+        //   displayDetails: false,
+        //   displayCompleted: true,
+        //   ongoing: [
+        //     {
+        //       name: 'CALS Credits',
+        //       type: 'Credits',
+        //       fulfilled: 12,
+        //       required: 55,
+        //       display: false
+        //     },
+        //     {
+        //       name: 'PE Credits',
+        //       type: 'Credits',
+        //       fulfilled: 1,
+        //       required: 2,
+        //       display: false
+        //     }
+        //   ],
+        //   completed: [
+        //     {
+        //       name: 'Quantitative Literacy',
+        //       type: 'Credits',
+        //       fulfilled: 2,
+        //       required: 2,
+        //       display: false
+        //     }
+        //   ]
+        // }
+        // {
+        //   name: 'COLLEGE REQUIREMENTS',
+        //   type: 'COLLEGE',
+        //   type: 'Credits',
+        //   progress: 46,
+        //   total: 120,
+        //   color: '2BBCC6',
+        //   displayDetails: false,
+        //   ongoing: [
+        //     {
+        //       name: 'CALS Credits',
+        //       type: 'Credits',
+        //       progress: 12,
+        //       total: 55,
+        //       additonalCourses: ['PE Courses', 'Swimming Requirement'],
+        //       satisfiableCourses: ['ice skating', 'bowling'],
+        //       display: false
+        //     },
+        //     {
+        //       name: 'PE Credits',
+        //       type: 'Credits',
+        //       progress: 1,
+        //       total: 2,
+        //       additonalCourses: ['PE Courses', 'Swimming Requirement'],
+        //       satisfiableCourses: ['ice skating', 'bowling'],
+        //       display: false
+        //     }
+        //   ],
+        //   displayCompleted: true,
+        //   completed: [
+        //     {
+        //       name: 'Quantitative Literacy',
+        //       type: 'Credits',
+        //       progress: 2,
+        //       total: 2,
+        //       display: false
+        //     },
+        //     {
+        //       name: 'Chemistry/Physics',
+        //       type: 'Credits',
+        //       progress: 1,
+        //       total: 2,
+        //       display: false
+        //     }
+        //   ]
+        // },
+        // {
+        //   name: 'MAJOR REQUIREMENTS',
+        //   type: 'MAJOR',
+        //   type: 'Courses',
+        //   progress: 4,
+        //   total: 8,
+        //   color: '508197',
+        //   displayDetails: false,
+        //   ongoing: [
+        //     {
+        //       name: 'CALS Credits',
+        //       type: 'Credits',
+        //       progress: 12,
+        //       total: 55,
+        //       display: false
+        //     },
+        //     {
+        //       name: 'PE ',
+        //       type: 'Credits',
+        //       progress: 1,
+        //       total: 2,
+        //       display: false
+        //     }
+        //   ],
+        //   displayCompleted: true,
+        //   completed: [
+        //     {
+        //       name: 'Quantitative Literacy',
+        //       type: 'Credits',
+        //       progress: 2,
+        //       total: 2,
+        //       display: false
+        //     },
+        //     {
+        //       name: 'Chemistry/Physics',
+        //       type: 'Credits',
+        //       progress: 2,
+        //       total: 2,
+        //       display: false
+        //     }
+        //   ]
+        // },
 
-        {
-          name: 'MINOR REQUIREMENTS',
-          type: 'MINOR',
-          count: 'Courses',
-          progress: 4,
-          total: 8,
-          color: '92C3E6',
-          displayDetails: false,
-          ongoing: [
-            {
-              name: 'CALS Credits',
-              count: 'Credits',
-              progress: 12,
-              total: 55,
-              display: false
-            },
-            {
-              name: 'PE ',
-              count: 'Credits',
-              progress: 1,
-              total: 2,
-              display: false
-            }
-          ],
-          displayCompleted: true,
-          completed: [
-            {
-              name: 'Quantitative Literacy',
-              count: 'Credits',
-              progress: 2,
-              total: 2,
-              display: false
-            },
-            {
-              name: 'Chemistry/Physics',
-              count: 'Credits',
-              progress: 1,
-              total: 2,
-              display: false
-            }
-          ]
-        }
-      ],
-
-      courses: {
-        id: randomId,
-        subject: 'PHIL',
-        code: 1100,
-        name: 'Introduction to Philosophy',
-        credits: 3,
-        semesters: ['Fall', 'Spring'],
-        color: '2BBCC6',
-        check: true,
-        requirementsMap: null
-      }
+        // {
+        //   name: 'MINOR REQUIREMENTS',
+        //   type: 'MINOR',
+        //   type: 'Courses',
+        //   progress: 4,
+        //   total: 8,
+        //   color: '92C3E6',
+        //   displayDetails: false,
+        //   ongoing: [
+        //     {
+        //       name: 'CALS Credits',
+        //       type: 'Credits',
+        //       progress: 12,
+        //       total: 55,
+        //       display: false
+        //     },
+        //     {
+        //       name: 'PE ',
+        //       type: 'Credits',
+        //       progress: 1,
+        //       total: 2,
+        //       display: false
+        //     }
+        //   ],
+        //   displayCompleted: true,
+        //   completed: [
+        //     {
+        //       name: 'Quantitative Literacy',
+        //       type: 'Credits',
+        //       progress: 2,
+        //       total: 2,
+        //       display: false
+        //     },
+        //     {
+        //       name: 'Chemistry/Physics',
+        //       type: 'Credits',
+        //       progress: 1,
+        //       total: 2,
+        //       display: false
+        //     }
+        //   ]
+        // }
+      ]
     };
   },
 
   methods: {
-
     turnDetails(index, bool) {
       this.reqs[index].displayDetails = !this.reqs[index].displayDetails;
     },
+
     turnSubDetails(index, id, bool) {
-      this.reqs[index].ongoing[id].display = !this.reqs[index].ongoing[id]
-        .display;
+      this.reqs[index].ongoing[id].display = !this.reqs[index].ongoing[id].display;
     },
 
     turnCompleted(index, bool) {
       this.reqs[index].displayCompleted = bool;
     },
 
-    getReqs() {
-      console.log("Todo check requirements");
+    async getReqs(coursesTaken, college, major) {
+
+      return getRequirements(coursesTaken, college, major);
+
+      async function getRequirements(coursesTaken, college, major) { // isTransfer = false
+        // TODO: make it so that it takes in classes corresponding with years/semesters for most accurate information
+        const coursesTakenWithInfo = {};
+        const courseData = await Promise.all(
+          coursesTaken.map(courseTaken => getCourseInfo(courseTaken))
+        );
+
+        for (let i = 0; i < coursesTaken.length; i += 1) { coursesTakenWithInfo[coursesTaken[i]] = courseData[i]; }
+
+        // prepare final output JSONs
+        let finalRequirementJSONs = [];
+
+        // PART 1: check university requirements
+        if (!reqsData.university) throw new Error('University requirements not found.');
+        const universityReqs = reqsData.university;
+        finalRequirementJSONs.push({groupName: "University", 
+          reqs: await iterateThroughRequirements(coursesTakenWithInfo, universityReqs.requirements, 'university')
+        });
+
+        // PART 2: check college requirements
+        if (!(college in reqsData.college)) throw new Error('College not found.');
+        const collegeReqs = reqsData.college[college];
+        finalRequirementJSONs.push({groupName: "College",
+          reqs: await iterateThroughRequirements(coursesTakenWithInfo, collegeReqs.requirements, 'college')
+        });
+
+        // PART 3: check major reqs
+        if (!(major in reqsData.major)) throw new Error('Major not found.');
+        const majorReqs = reqsData.major[major];
+        finalRequirementJSONs.push({groupName: "Major",
+          reqs: await iterateThroughRequirements(coursesTakenWithInfo, majorReqs.requirements, 'major')
+        });
+
+        return finalRequirementJSONs;
+      }
+
+
+      /**
+       * Loops through requirement data and compare all courses on (to identify whether they satisfy the requirement)
+       * @param {*} allCoursesTakenWithInfo : object of courses taken with API information (CS 2110: {info})
+       * @param {*} allRequirements : requirements in requirements format from reqs.json (college, major, or university requirements)
+       * @param {*} requirementType : type of requirement being checked (college, major, or university)
+       */
+      async function iterateThroughRequirements(allCoursesTakenWithInfo, allRequirements, requirementType) {
+        // array of requirement status information to be returned
+        const requirementJSONs = [];
+
+        // helper to recursively call when an object has subpaths
+        function helper(coursesTakenWithInfo, requirements, rType, parentName = null) {
+          for (const requirement of requirements) {
+            // if(!isTransfer && requirement.applies === "transfers") continue;
+            // temporarily skip these until we can implement them later
+
+            // Recursively call function if there are subpaths
+            if (requirement.multiplePaths) {
+              const requirementName = requirement.name;
+              requirementJSONs.push({ name: requirementName, paths: [], isComplete: false });
+              helper(coursesTakenWithInfo, requirement.paths, requirementName);
+              continue;
+            }
+
+            let totalRequirementCredits = 0;
+            let totalRequirementCount = 0;
+            const coursesThatFulilledRequirement = [];
+            // check each course to see if it fulfilled that requirement
+
+            const codes = Object.keys(coursesTakenWithInfo);
+
+            // If not in path, push new object to requirementsJSONs
+            for (const code of codes) {
+              const courseInfo = coursesTakenWithInfo[code];
+
+              const indexIsFulfilled = checkIfCourseFulfilled(courseInfo, requirement.search, requirement.includes);
+
+              if (indexIsFulfilled) {
+                // depending on what it is fulfilled by, either increase the count or credits you took
+                switch (requirement.fulfilledBy) {
+                  case 'courses':
+                    totalRequirementCount += 1;
+                    break;
+                  case 'credits':
+                    totalRequirementCredits += courseInfo.enrollGroups[0].unitsMaximum;
+                    break;
+                  case 'self-check':
+                    continue;
+                  default:
+                    throw new Error('Fulfillment type unknown.');
+                }
+
+                // add the course to the list of courses used to fulfill that one requirement
+                coursesThatFulilledRequirement.push(code);
+              }
+            }
+
+            const generatedResults = createRequirementJSON(requirement, totalRequirementCredits, totalRequirementCount, coursesThatFulilledRequirement, );
+
+            // If at end path (no parent path)
+            if (!parentName) requirementJSONs.push(generatedResults);
+            // If in path, append to path of parent
+            else {
+              const parent = requirementJSONs.find(key => key.name === parentName);
+              parent.paths.push(generatedResults);
+              parent.isComplete = parent.isComplete || generatedResults.isComplete;
+            }
+          }
+        }
+
+        helper(allCoursesTakenWithInfo, allRequirements, requirementType);
+
+        return requirementJSONs;
+      }
+
+      /**
+       * Creates results in object format from information
+       * @param {*} requirement : the requirement information as object
+       * @param {*} totalRequirementCredits : total credits of courses that satisfied requirement
+       * @param {*} totalRequirementCount : total number of courses that satisfied requirement
+       * @param {*} coursesThatFulilledRequirement : courses that satisfied requirement
+       */
+      function createRequirementJSON(requirement, totalRequirementCredits, totalRequirementCount, coursesThatFulilledRequirement, requirementType) {
+        const requirementFulfillmentData = {
+          name: requirement.name,
+          type: requirement.fulfilledBy,
+          courses: coursesThatFulilledRequirement,
+          required: requirement.minCount
+        };
+        let fulfilled;
+        switch (requirement.fulfilledBy) {
+          case 'courses':
+            fulfilled = totalRequirementCount;
+            break;
+          case 'credits':
+            fulfilled = totalRequirementCredits;
+            break;
+          case 'self-check':
+            fulfilled = null;
+            break;
+          default:
+            throw new Error('Fulfillment type unknown.');
+        }
+
+        requirementFulfillmentData.fulfilled = fulfilled;
+
+        if (requirement.progressBar) requirementFulfillmentData.progressBar = true;
+        return requirementFulfillmentData;
+      }
+
+      /**
+       * Given a course abbreviation (i.e. INFO 1300), it will split it up into the subject and number, returned as a dictionary
+       * (i.e. INFO 1300 => {"subject" : INFO, "courseNumber" : 1300})
+       *
+       * @return the number of credits the course is worth
+       */
+      function parseCourseAbbreviation(courseAbbreviation) {
+        const regex = /([a-zA-Z]+) ([0-9][0-9][0-9][0-9]$)?/g;
+        const matches = regex.exec(courseAbbreviation);
+        if (matches === null) throw new Error('Invalid course abbreviation');
+        return { subject: matches[1].toUpperCase(), courseNumber: matches[2] };
+      }
+
+      /**
+       * Given a course code and a roster, get all course info from Cornell API
+       * @param {*} courseCode : code name of the course to search (CS 2110)
+       * @param {*} semester : the roster name to search from (FA19)
+       */
+      function getCourseInfo(courseCode) {
+        const courseAbbrev = parseCourseAbbreviation(courseCode);
+        const subject = courseAbbrev.subject.toUpperCase();
+        const number = courseAbbrev.courseNumber;
+
+        return new Promise((resolve, reject) => {
+          // Using Firebase
+          const coursesCollection = fb.db.collection('courses');
+          const courseRef = coursesCollection.doc(subject + number);
+
+          courseRef.get()
+            .then(doc => {
+              if (!doc.exists) reject(new Error('No document exists'));
+              else resolve(doc.data());
+            })
+            .catch(() => {
+              reject(new Error('An error occured.'));
+            });
+        });
+      }
+
+      /**
+       * Check if a code matches the course name (CS 2110 and CS 2*** returns true, AEM 3110 and AEM 32** returns false)
+       * @param {*} courseName : name of the course (as a code)
+       * @param {*} code : code to check courseName (can contain * to denote any value)
+       */
+      function ifCodeMatch(courseName, code) {
+        for (let i = 0; i < courseName.length; i += 1) {
+          if (code[i] !== '*' && courseName[i] !== code[i]) return false;
+        }
+
+        return true;
+      }
+
+      /**
+       * Check if the course fullfills the given requirement. Returns true if fulfills requirement. False otherswise
+       * @param {*} courseInfo : information of the course from API data
+       * @param {*} search : the scope of search for the requirement (e.g all-eligible, code, catalogDistr)
+       * @param {*} includes : the query for the search (e.g (MQR-AS), CS 2***)
+       */
+      function checkIfCourseFulfilled(courseInfo, search, includes) {
+        if (search === 'all' || search === 'all-eligible' || search === 'self-check') return true;
+        for (const [i, include] of includes.entries()) {
+          for (const option of include) {
+            if (search === 'code') {
+              if (ifCodeMatch(`${courseInfo.subject} ${courseInfo.catalogNbr}`, option)) {
+                // Important: removes array option list from requirements
+                if (includes.length > 1) includes.splice(i, 1);
+                return true;
+              }
+            } else if (courseInfo[search].includes(option)) return true;
+          }
+        }
+
+        return false;
+      }
     }
   }
 };
@@ -395,8 +675,8 @@ export default {
 }
 
 .depth-req {
-  margin: 0.5rem 0 0.5rem 0;
-  height: 14px;
+  margin: 0.5rem 0 0.1rem 0;
+  min-height: 14px;
 }
 
 .sub-req-div{
